requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1

langchain==0.1.16
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.1.3

faiss-cpu==1.8.0.post1
tavily-python==0.3.3

groq==0.8.0
httpx==0.27.0

pypdf==4.2.0
wikipedia==1.4.0
beautifulsoup4==4.12.3

numpy==1.26.4
pandas==2.2.2
